refactor[package] : 패키지 구조 변경 및 환경 설정
