+++ src/main/java/com/example/shortudy/domain/keyword/dto/request/TagRequest.java
@@ -1,4 +1,4 @@
-package com.example.shortudy.domain.tag.dto.request;
+package com.example.shortudy.domain.keyword.dto.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
