+++ src/main/java/com/example/shortudy/domain/keyword/repository/TagRepository.java
@@ -1,6 +1,6 @@
-package com.example.shortudy.domain.tag.repository;
+package com.example.shortudy.domain.keyword.repository;
 
-import com.example.shortudy.domain.tag.entity.Tag;
+import com.example.shortudy.domain.keyword.entity.Tag;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
