+++ src/main/java/com/example/shortudy/domain/shorts/service/ShortsService.java
@@ -8,8 +8,8 @@ import com.example.shortudy.domain.shorts.dto.ShortsUploadRequest;
 import com.example.shortudy.domain.shorts.entity.Shorts;
 import com.example.shortudy.domain.shorts.entity.ShortsStatus;
 import com.example.shortudy.domain.shorts.repository.ShortsRepository;
-import com.example.shortudy.domain.tag.entity.Tag;
-import com.example.shortudy.domain.tag.service.TagService;
+import com.example.shortudy.domain.keyword.entity.Tag;
+import com.example.shortudy.domain.keyword.service.TagService;
 import com.example.shortudy.domain.user.entity.User;
 import com.example.shortudy.domain.user.repository.UserRepository;
 import org.springframework.data.domain.Page;
