+++ src/main/java/com/example/shortudy/domain/user/service/AuthServiceImpl.java
@@ -8,9 +8,6 @@ import com.example.shortudy.domain.user.dto.response.AuthStatusResponse;
 import com.example.shortudy.domain.user.dto.response.UserLoginResponse;
 import com.example.shortudy.domain.user.entity.User;
 import com.example.shortudy.domain.user.repository.UserRepository;
-import com.example.shortudy.global.error.exception.EmailAlreadyExistsException;
-import com.example.shortudy.global.error.exception.InvalidPasswordException;
-import com.example.shortudy.global.error.exception.UserNotFoundException;
 import com.example.shortudy.global.jwt.JwtTokenProvider;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
