+++ src/main/java/com/example/shortudy/domain/user/service/UserService.java
@@ -5,7 +5,6 @@ import com.example.shortudy.domain.shorts.repository.ShortsRepository;
 import com.example.shortudy.domain.user.dto.UserResponse;
 import com.example.shortudy.domain.user.entity.User;
 import com.example.shortudy.domain.user.repository.UserRepository;
-import com.example.shortudy.global.error.exception.UserNotFoundException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
