+++ src/main/java/com/example/shortudy/global/error/ErrorCode.java
@@ -1,0 +1,11 @@
+package com.example.shortudy.global.error;
+
+public enum ErrorCode {
+    UserNotFoundException,
+    ResourceNotFoundException,
+    InvalidPasswordException,
+    EmailAlreadyExistsException
+
+
+
+}
